Cleaned up code and refactored highlighting code
