+++ src/components/AppBody.jsx
@@ -43,7 +43,7 @@ class AppBody extends React.Component {
   }
 
   onGutterClick(codeMirrorInstance, lineNumber) {
-    this.snippetArea.emphasizeLine(this.state.snippet, lineNumber);
+    // this.snippetArea.emphasizeLine(this.state.snippet, lineNumber);
     if (this.state.annotations[String(lineNumber)] === undefined) {
       this.setState({
         annotationDisplay: 'create',
@@ -70,7 +70,6 @@ class AppBody extends React.Component {
   }
 
   closeAnnotation() {
-    this.snippetArea.doHighlighting();
     this.snippetArea.triggerHighlight(
       this.state.snippet, 
       this.state.AST, 
@@ -211,6 +210,7 @@ class AppBody extends React.Component {
               switchReadOnlyMode={this.switchReadOnlyMode}
               title={this.state.snippetTitle}
               filters={this.state.filters}
+              emphasizeLine={this.state.annotationDisplayProps.lineNumber}
             />
           </div>
           <div className="col-md-4">
